requirements.txt
cdp-langchain
langchain-openai
langgraph
web3>=6.0.0
pyyaml
python-dotenv
torch
numpy
pandas
scikit-learn
textblob
vaderSentiment
eth-account
eth-typing
eth-utils
tweepy
aiohttp
cryptography